updated error messaging and controllers to never allow actions to unlogged in users

+++ app/controllers/works_controller.rb
@@ -17,15 +17,22 @@ class WorksController < ApplicationController
   end
 
   def create
-    @work = Work.new(work_params)
-    @work.vote_count = 0
-    
-    if @work.save
-      redirect_to work_path(@work.id)
-      flash[:success] = "#{@work.title} added successfully"
+    user = session[:user_id]
+    if user
+      @work = Work.new(work_params)
+      @work.vote_count = 0
+      
+      if @work.save
+        redirect_to work_path(@work.id)
+        flash[:success] = "#{@work.title} added successfully!"
+      else
+        flash.now[:error] = "Your new work was not added successfully. Please fix errors before adding."
+        render new_work_path
+        return
+      end
     else
-      flash.now[:error] = @work.errors.messages
-      render new_work_path
+      flash[:error] = "Please login first to add new work."
+      redirect_to login_path
       return
     end
   end
@@ -38,32 +45,46 @@ class WorksController < ApplicationController
   end
 
   def update
-    if @work.nil?
-      head :not_found
-      return
-    elsif @work.update(work_params)
-      flash[:success] = "#{@work.title} updated successfully"
-      redirect_to work_path(@work.id)
-      return
+    user = session[:user_id]
+    if user
+      if @work.nil?
+        head :not_found
+        return
+      elsif @work.update(work_params)
+        flash[:success] = "#{@work.title} updated successfully"
+        redirect_to work_path(@work.id)
+        return
+      else
+        flash[:error] = "Work NOT updated successfully"
+        redirect_back(fallback_location: works_path)
+        return
+      end
     else
-      flash.now[:error] = "Work NOT updated successfully"
-      render edit_work_path
+      flash[:error] = "Please login before you can update works."
+      redirect_to login_path
       return
     end
   end
 
   def destroy
-    work = Work.find_by(id: params[:id])
-    if work.nil?
-      head :not_found
-      return
-    elsif work.destroy
-      flash[:success] = "#{work.title} deleted successfully"
-      redirect_to works_path
-      return
+    user = session[:user_id]
+    if user
+      work = Work.find_by(id: params[:id])
+      if work.nil?
+        head :not_found
+        return
+      elsif work.destroy
+        flash[:success] = "#{work.title} deleted successfully"
+        redirect_to works_path
+        return
+      else
+        flash[:error] = "Work NOT deleted successfully"
+        redirect_back(fallback_location: root_path)
+      end
     else
-      flash[:error] = "Work NOT deleted successfully"
-      redirect_back(fallback_location: root_path)
+      flash[:error] = "Please login before you can delete works."
+      redirect_to login_path
+      return
     end
   end
